refactoring, update argocd plugins
