+++ images/argocd-custom-image/terra-helmfile-argocd/terra-helmfile-argocd.sh
@@ -12,8 +12,9 @@ set -x
 if [[ "$1" == 'init' ]]; then
   : # Nothing to do
 elif [[ "$1" == 'generate' ]]; then
-  # Delegate to render script
-  export TERRA_HELMFILE_PATH=$( pwd )
+  # Delegate to `thelma render`
+  THELMA_HOME=$( pwd )
+  export THELMA_HOME
   thelma render --stdout --argocd
 else
   echo "Usage: ${0} (init|generate)" >&2
