+++ tools/internal/shell/shellmock/example_test.go
