+++ tools/internal/thelma/charts/mirror/mirror_test.go
@@ -2,7 +2,7 @@ package mirror
 
 import (
 	"github.com/broadinstitute/terra-helmfile-images/tools/internal/shell"
-	"github.com/broadinstitute/terra-helmfile-images/tools/internal/shellmock"
+	"github.com/broadinstitute/terra-helmfile-images/tools/internal/shell/shellmock"
 	"github.com/broadinstitute/terra-helmfile-images/tools/internal/thelma/charts/publish"
 	"github.com/broadinstitute/terra-helmfile-images/tools/internal/thelma/charts/repo/index"
 	"github.com/broadinstitute/terra-helmfile-images/tools/internal/thelma/tools/helm"
