+++ tools/internal/thelma/charts/publish/publisher_test.go
@@ -3,7 +3,7 @@ package publish
 import (
 	"fmt"
 	"github.com/broadinstitute/terra-helmfile-images/tools/internal/shell"
-	"github.com/broadinstitute/terra-helmfile-images/tools/internal/shellmock"
+	"github.com/broadinstitute/terra-helmfile-images/tools/internal/shell/shellmock"
 	"github.com/broadinstitute/terra-helmfile-images/tools/internal/thelma/charts/repo"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
