+++ tools/internal/thelma/render/helmfile/helmfile_test.go
@@ -2,7 +2,7 @@ package helmfile
 
 import (
 	"github.com/broadinstitute/terra-helmfile-images/tools/internal/shell"
-	"github.com/broadinstitute/terra-helmfile-images/tools/internal/shellmock"
+	"github.com/broadinstitute/terra-helmfile-images/tools/internal/shell/shellmock"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
